# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/illustrations": {
            "get": {
                "description": "Get a paginated list of illustrations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "illustrations",
                    "admin"
                ],
                "summary": "Get All Illustrations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.PaginatedIllustrationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new illustration (Admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "illustrations"
                ],
                "summary": "Create Illustration",
                "parameters": [
                    {
                        "description": "Illustration Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.CreateIllustrationReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.IllustrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/illustrations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific illustration by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "illustrations"
                ],
                "summary": "Get Illustration by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Illustration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.IllustrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing illustration (Admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "illustrations"
                ],
                "summary": "Update Illustration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Illustration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Illustration Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.UpdateIllustrationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.IllustrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an illustration by ID (Admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "illustrations"
                ],
                "summary": "Delete Illustration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Illustration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/illustrations": {
            "get": {
                "description": "Get a paginated list of illustrations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "illustrations",
                    "admin"
                ],
                "summary": "Get All Illustrations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.PaginatedIllustrationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/illustrations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific illustration by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "illustrations"
                ],
                "summary": "Get Illustration by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Illustration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_illustration.IllustrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gotickets_internal_httpresponse.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_domain_illustration.CreateIllustrationReq": {
            "type": "object",
            "required": [
                "contentText"
            ],
            "properties": {
                "contentText": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "internal_domain_illustration.IllustrationResponse": {
            "type": "object",
            "properties": {
                "contentText": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "internal_domain_illustration.PaginatedIllustrationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_illustration.IllustrationResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_illustration.UpdateIllustrationReq": {
            "type": "object",
            "required": [
                "contentText"
            ],
            "properties": {
                "contentText": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
...
    }
}